fix(dataSets):adapted component to location group's data sets
